Request > Return `Session` instead of `SessionInterface`
Symfony 6 removed the `Session` and `FlashBagInterface` services.

That means that the only way to get the FlashBag (to add flashes) it to use `$request->getSession()->getFlashBag()`.

But the problem is, `getFlashBag` is not defined on the `SessionInterface`:
```
Call to an undefined method Symfony\Component\HttpFoundation\Session\SessionInterface::getFlashBag().
```

To make everyone's life easier, let's change the return type to `Sesssion` instead.
This is what's returned in a default Symfony application.