make creation of ServiceMap and ParameterMap lazy
Currently one can not use PHPStans `bootstrapFiles` to create the
container (dump) which could then be ready by
`symfony.containerXmlPath`. This as the order of initialization results
in the `XmlServiceMapFactory` / `XmlParameterMapFactory` `create`
methods being called before the `bootstrapFiles` are executed. By making
the `ServiceMap` / `ParameterMap` lazy these will only be created when
actually used, which is after the `bootstrapFiles` have been executed.